Fix lint and refactor
cc @Mastercuber
Send out an activity should be the responsibility of the middleware.
