+++ src/activitypub/utils/index.js
@@ -75,7 +75,9 @@ export function signAndSend (activity, fromName, targetDomain, url) {
         headers: {
           'Host': targetDomain,
           'Date': date,
-          'Signature': createSignature({privateKey, keyId: `http://${activityPub.domain}/activitypub/users/${fromName}#main-key`, url,
+          'Signature': createSignature({ privateKey,
+            keyId: `http://${activityPub.domain}/activitypub/users/${fromName}#main-key`,
+            url,
             headers: {
               'Host': targetDomain,
               'Date': date,
